chore: rename hooks from scaffold-stylus to scaffold-eth
