+++ packages/nextjs/app/blockexplorer/_components/AddressLogsTab.tsx
@@ -1,5 +1,5 @@
 import { Address } from "viem";
-import { useContractLogs } from "~~/hooks/scaffold-stylus";
+import { useContractLogs } from "~~/hooks/scaffold-eth";
 import { replacer } from "~~/utils/scaffold-stylus/common";
 
 export const AddressLogsTab = ({ address }: { address: Address }) => {
