+++ packages/nextjs/app/blockexplorer/page.tsx
@@ -3,8 +3,8 @@
 import { useEffect, useState } from "react";
 import { PaginationButton, TransactionsTable } from "./_components";
 import type { NextPage } from "next";
-import { useFetchBlocks } from "~~/hooks/scaffold-stylus";
-import { useTargetNetwork } from "~~/hooks/scaffold-stylus/useTargetNetwork";
+import { useFetchBlocks } from "~~/hooks/scaffold-eth";
+import { useTargetNetwork } from "~~/hooks/scaffold-eth/useTargetNetwork";
 import { notification } from "~~/utils/scaffold-stylus";
 import { arbitrumNitro } from "~~/utils/scaffold-stylus/chain";
