+++ packages/nextjs/app/debug/_components/contract/DisplayVariable.tsx
@@ -7,7 +7,7 @@ import { Abi, AbiFunction } from "abitype";
 import { Address } from "viem";
 import { useContractRead } from "wagmi";
 import { ArrowPathIcon } from "@heroicons/react/24/outline";
-import { useAnimationConfig } from "~~/hooks/scaffold-stylus";
+import { useAnimationConfig } from "~~/hooks/scaffold-eth";
 import { notification } from "~~/utils/scaffold-stylus";
 
 type DisplayVariableProps = {
