+++ packages/nextjs/components/ScaffoldEthAppWithProviders.tsx
@@ -9,7 +9,7 @@ import { Footer } from "~~/components/Footer";
 import { Header } from "~~/components/Header";
 import { BlockieAvatar } from "~~/components/scaffold-stylus";
 import { ProgressBar } from "~~/components/scaffold-stylus/ProgressBar";
-import { useNativeCurrencyPrice } from "~~/hooks/scaffold-stylus";
+import { useNativeCurrencyPrice } from "~~/hooks/scaffold-eth";
 import { useGlobalState } from "~~/services/store/store";
 import { wagmiConfig } from "~~/services/web3/wagmiConfig";
 import { appChains } from "~~/services/web3/wagmiConnectors";
