+++ packages/nextjs/components/scaffold-stylus/FaucetButton.tsx
@@ -6,7 +6,7 @@ import { privateKeyToAccount } from "viem/accounts";
 import { useAccount, useNetwork } from "wagmi";
 import { useBalance } from "wagmi";
 import { BanknotesIcon } from "@heroicons/react/24/outline";
-import { useTransactor } from "~~/hooks/scaffold-stylus";
+import { useTransactor } from "~~/hooks/scaffold-eth";
 import { arbitrumNitro } from "~~/utils/scaffold-stylus/chain";
 
 // Number of ETH faucet sends to an address
