+++ packages/nextjs/components/scaffold-stylus/RainbowKitCustomConnectButton/index.tsx
@@ -10,8 +10,8 @@ import { WrongNetworkDropdown } from "./WrongNetworkDropdown";
 import { ConnectButton } from "@rainbow-me/rainbowkit";
 import { Address } from "viem";
 import { useConnect } from "wagmi";
-import { useAutoConnect, useNetworkColor } from "~~/hooks/scaffold-stylus";
-import { useTargetNetwork } from "~~/hooks/scaffold-stylus/useTargetNetwork";
+import { useAutoConnect, useNetworkColor } from "~~/hooks/scaffold-eth";
+import { useTargetNetwork } from "~~/hooks/scaffold-eth/useTargetNetwork";
 import scaffoldConfig from "~~/scaffold.config";
 import { burnerWalletConfig } from "~~/services/web3/wagmi-burner/burnerWalletConfig";
 import { getBlockExplorerAddressLink } from "~~/utils/scaffold-stylus";
