+++ packages/nextjs/hooks/scaffold-eth/useAnimationConfig.ts
