+++ packages/nextjs/hooks/scaffold-eth/useAutoConnect.ts
