+++ packages/nextjs/hooks/scaffold-eth/useBurnerWallet.ts
