+++ packages/nextjs/hooks/scaffold-eth/useDeployedContractInfo.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { useTargetNetwork } from "../scaffold-stylus/useTargetNetwork";
+import { useTargetNetwork } from "./useTargetNetwork";
 import { useIsMounted } from "usehooks-ts";
 import { usePublicClient } from "wagmi";
 import { Contract, ContractCodeStatus, ContractName, contracts } from "~~/utils/scaffold-stylus/contract";
