+++ packages/nextjs/hooks/scaffold-eth/useOutsideClick.ts
