--- packages/nextjs/hooks/scaffold-stylus/index.ts
@@ -1,15 +1,0 @@
-export * from "./useAnimationConfig";
-export * from "../scaffold-stylus/useBurnerWallet";
-export * from "../scaffold-stylus/useDeployedContractInfo";
-export * from "../scaffold-stylus/useNativeCurrencyPrice";
-export * from "../scaffold-stylus/useNetworkColor";
-export * from "../scaffold-stylus/useOutsideClick";
-export * from "../scaffold-stylus/useScaffoldContract";
-export * from "../scaffold-stylus/useScaffoldContractRead";
-export * from "../scaffold-stylus/useScaffoldContractWrite";
-export * from "../scaffold-stylus/useScaffoldEventSubscriber";
-export * from "../scaffold-stylus/useScaffoldEventHistory";
-export * from "../scaffold-stylus/useTransactor";
-export * from "../scaffold-stylus/useFetchBlocks";
-export * from "../scaffold-stylus/useContractLogs";
-export * from "../scaffold-stylus/useAutoConnect";
